% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gobmx.R
\docType{package}
\name{gobmx}
\alias{gobmx}
\title{gobmx}
\description{
list of palettes inspired in the ones by the Mexican government.
This is not an official product.
}
\details{
list of palettes inspired in the ones by the Mexican government
}
