% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{gobmx_palettes}
\alias{gobmx_palettes}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 6.
}
\usage{
gobmx_palettes
}
\description{
Use \code{\link{gobmx_palette}} to construct palettes of desired length.
}
\keyword{datasets}
